+++ contracts/Vaults/FujiVault.sol
@@ -59,8 +59,8 @@ contract FujiVault is IVault, VaultBaseUpgradeable, ReentrancyGuardUpgradeable {
 
   string public name;
 
-  uint256 internal _collateralAssetDecimals;
-  uint256 internal _borrowAssetDecimals;
+  uint8 internal _collateralAssetDecimals;
+  uint8 internal _borrowAssetDecimals;
 
   modifier isAuthorized() {
     require(
@@ -98,7 +98,7 @@ contract FujiVault is IVault, VaultBaseUpgradeable, ReentrancyGuardUpgradeable {
       _collateralAssetDecimals = 18;
     } else {
       collateralSymbol = IERC20Extended(_collateralAsset).symbol();
-      _collateralAssetDecimals = uint256(IERC20Extended(_collateralAsset).decimals());
+      _collateralAssetDecimals = IERC20Extended(_collateralAsset).decimals();
     }
 
     if (_borrowAsset == ETH) {
@@ -106,7 +106,7 @@ contract FujiVault is IVault, VaultBaseUpgradeable, ReentrancyGuardUpgradeable {
       _borrowAssetDecimals = 18;
     } else {
       borrowSymbol = IERC20Extended(_borrowAsset).symbol();
-      _borrowAssetDecimals = uint256(IERC20Extended(_borrowAsset).decimals());
+      _borrowAssetDecimals = IERC20Extended(_borrowAsset).decimals();
     }
 
     name = string(abi.encodePacked("Vault", collateralSymbol, borrowSymbol));
@@ -511,7 +511,7 @@ contract FujiVault is IVault, VaultBaseUpgradeable, ReentrancyGuardUpgradeable {
       vAssets.borrowAsset,
       _collateralAssetDecimals
     );
-    uint256 minimumReq = (_amount * price) / (10**_borrowAssetDecimals);
+    uint256 minimumReq = (_amount * price) / (10**uint256(_borrowAssetDecimals));
     if (_withFactors) {
       return (minimumReq * (collatF.a) * (safetyF.a)) / (collatF.b) / (safetyF.b);
     } else {
